requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
discord.py==2.5.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pefile==2023.2.7
propcache==0.3.2
Pygments==2.19.2
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
